VNPay tested successfully

+++ Game_Web/server/server/src/main/java/com/example/server/DTO/CartResponseDTO.java
@@ -2,12 +2,14 @@ package com.example.server.DTO;
 
 import lombok.*;
 
+import java.math.BigDecimal;
+
 @AllArgsConstructor
 @NoArgsConstructor
 @Getter
 @Setter
 @Builder
 public class CartResponseDTO {
-    private Long totalMoney;
+    private BigDecimal totalMoney;
     private Long totalQuantity;
 }
